refactor(ai-plan-import): renomme applyConsolidations en updateActionsWithConsolidatedEntries
